% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.login.replacement.R
\name{datashield.login}
\alias{datashield.login}
\title{Extends datashield.login}
\usage{
datashield.login(...)
}
\arguments{
\item{...}{the parameters for datashield.login (logins dataframe, etc).
See the documentation for that function for details.}
}
\value{
a vector containing the names of all the establised connections (real and fake)
}
\description{
Executes the login then redefines the data.frame function and sets the name of the server in the respective remote sessions
}
\details{
Some functions in the server need  modifications, namely:
* The remote server session needs to know its own server name. This functions sets this name in a hidden object called .whoami
* Datashield and dssSwissKnife need strings (as well as dates) loaded as factors and not chars (or dates). This function redefines
the data.frame function on all the remote sessions in order to achieve this.
}
