% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToInsulinRhapsody.R
\name{ds2MonotherapyNoInsulin}
\alias{ds2MonotherapyNoInsulin}
\title{Monotherapy no insulin}
\usage{
ds2MonotherapyNoInsulin(
  symbol,
  treatment.df,
  diag.df,
  insulin.treatments = list(CMCAT = "INSULINS AND ANALOGUES"),
  trt.date.col = "CMDTC",
  diag.date.col = "MHDTC",
  occur.col = list(CMOCCUR = list(yes = "Y", no = "N")),
  diag.occur.col = list(MHOCCUR = list(yes = "Y", no = "N")),
  diag.visit = "DIAGNOSIS",
  diagnosis = list(MHTERM = "TYPE 2 DIABETES"),
  years.after.diagnosis = 5,
  days.offset = 30,
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{symbol}{the name of the new dataframe}

\item{treatment.df}{the name of the treatment dataframe (CM table)}

\item{diag.df}{the name of the diagnosis dataframe (MH table)}

\item{insulin.treatments}{encoding of the insulin treatments to be excluded. A list with exactly one named element (the default name CMCAT is the treatment column name in treatment.df  and
should be correct in most cases ). The element must be a vector with one ore more treatments present in treatment.df$CMCAT (again, the default should suffice)}

\item{trt.date.col}{the name of the date column in treatment.df. The default (CMDTC) should be OK in all cohorts with the exception of dcs where the name is CMSTDTC}

\item{diag.date.col}{the name of the date column in diag.df. The default (MHDTC) should be OK in all cohorts.}

\item{occur.col}{a list with exactly one element whose name must match the name fo the occurence column in treatment.df ('CMOCCUR' should be fine everywhere).
The element must be a list with exactly 2 elements, the encoding of the 'yes' and 'no' values. Again the defaults should be fine.
If in doubt, ds.levels(treatment.df$CMOCCUR) or even better ds2.show.factors(treatment.df) should give the necessary info.}

\item{diag.occur.col}{a list exactly like the one above, this time with the column name and the levels from diag.df}

\item{diag.visit}{a character, what visit should we consider for diagnosis.}

\item{diagnosis}{a character, name of the diagnosis, the default should be fine for all cohorts}

\item{years.after.diagnosis}{a number what is the cutoff point? (default 5 years)}

\item{days.offset}{a number, how far to look around the cutoff point, in days, default one month}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\value{
It doesn't return anything,  it creates a filtered dataframe on the remote node with the columns SUBJID, DIFF_DAYS, CASE
(the CASE column is always 0 and it serves to further modeling with clogit)
}
\description{
Implements the controls for time to insulin, defined as a diet or monotherapy without insulin treatment at 5 years (default) from diagnosis
}
