% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToInsulinRhapsody.R
\name{dssTimeToInsulin}
\alias{dssTimeToInsulin}
\title{Time to insulin}
\usage{
dssTimeToInsulin(
  symbol,
  treatment.df,
  lab.df,
  diag.df,
  treatments = list(CMCAT = "INSULINS AND ANALOGUES"),
  trt.date.col = "CMDTC",
  lab.date.col = "LBDTC",
  diag.date.col = "MHDTC",
  occur.col = list(CMOCCUR = list(yes = "Y", no = "N")),
  diag.occur.col = list(MHOCCUR = list(yes = "Y", no = "N")),
  diag.visit = "DIAGNOSIS",
  diagnosis = list(MHTERM = "TYPE 2 DIABETES"),
  min.treatment.days = 180,
  labtest = list(LBTESTCD = "HBA1C"),
  min.measurement = list(LBORRES = 77),
  measure = list(LBORRESU = "mmol/m"),
  min.days.apart = 90,
  criteria = "both",
  async = TRUE,
  datasources = NULL
)
}
\arguments{
\item{symbol}{the name of the new dataframe}

\item{treatment.df}{the name of the treatment dataframe (CM table)}

\item{lab.df}{the name of the lab measures dataframe (LB table)}

\item{diag.df}{the name of the diagnosis dataframe (MH table)}

\item{treatments}{a list with exactly one named element (the default name CMCAT is the treatment column name in treatment.df  and
should be correct in most cases ). The element must be a vector with one ore more treatments present in treatment.df$CMCAT (again, the default should suffice)}

\item{trt.date.col}{the name of the date column in treatment.df. The default (CMDTC) should be OK in all cohorts with the exception of dcs where the name is CMSTDTC}

\item{lab.date.col}{the name of the date column in lab.df. The default (LBDTC) should be OK in all cohorts.}

\item{diag.date.col}{the name of the date column in diag.df. The default (MHDTC) should be OK in all cohorts.}

\item{occur.col}{a list with exactly one element whose name must match the name fo the occurence column in treatment.df ('CMOCCUR' should be fine everywhere).
The element must be a list with exactly 2 elements, the encoding of the 'yes' and 'no' values. Again the defaults should be fine.
If in doubt, ds.levels(treatment.df$CMOCCUR) or even better ds2.show.factors(treatment.df) should give the necessary info.}

\item{diag.occur.col}{a list exactly like the one above, this time with the column name and the levels from diag.df}

\item{diag.visit}{a character, what visit should we consider for diagnosis.}

\item{diagnosis}{a character, name of the diagnosis, the default should be fine for all cohorts}

\item{min.treatment.days}{how many days on insulin (criteria 1 above)}

\item{labtest}{a list with exactly one element containing the name of the column pointing to exactly one value - the name of the test (HBA1C in most cases)}

\item{min.measurement}{a list with exactly one element containing exactly one value: name of the column in lab.df pointing to the minimum value (default 77 ~ 8.5\% from the second criterion above)}

\item{measure}{a list with exactly one element containing exactly one value: name of the unit column in lab.df pointing to the unit used (default mmol/m)}

\item{min.days.apart}{an integer, how many days apart the measurements of HBA1C (second criterion above)}

\item{criteria}{a character in c('both', 'first', 'second'). Which of the criteria above should we apply?}

\item{async}{same as in datashield.assign}

\item{datasources}{same as in datashield.assign}
}
\value{
It doesn't return anything,  it creates a filtered dataframe on the remote node with the columns SUBJID, DAYS_TO_INSULIN, CASE
(the CASE column is always 1 and it serves to further modeling with clogit)
}
\description{
Implements the criterion for the clinical endpoint for time to insulin, namely the earlier of:

1) starting sustained (more than 6 months by default) insulin treatment

2) 2 or more HBA1C measurements higher than 8.5\% (default), more than 3 months apart (default) when on 2 or more non-insulin therapies
}
